requirements.txt
Flask
psycopg2-binary
python-dotenv
flask_cors